# src/docs/swagger.doc.json
{ 
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Match",
    "description": "Workout Match is an API for creating and managing customized exercises and workouts. With this API, you can create exercises with sets, reps, and exercise classification (type), as well as group them into workouts that can have a defined difficulty level. Additionally, Workout Match also offers the ability to search for exercises and workouts from other users and copy them to your own account, which is especially useful for those looking for new workout ideas and variations.",
    "contact": {
      "email": "arthurgbsf@gmail.com"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Database"
    },
    {
      "url": "https://workout-match.onrender.com/api-docs",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/exercises": {
      "get": {
        "parameters": [{
          "in":"query", 
          "name": "user", 
          "required": false,
          "schema":{"type": "boolean"},
        "description": "Optional. Pass 'user = true' to get the logged user exercises."}
        ],
        "description": "Route to get all users exercises registered. Exercises copied by an user not appear in this request.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Exercises"
                  }
                },
                "examples": {
                  "exercises": {
                    "value": [
                      {
                        "_id": "644a7d690ff85088d51fcb4c",
                        "exercise": "Push ups",
                        "sets": "5",
                        "reps": "15",
                        "type": "push",
                        "createdBy": "644a7d0d0ff85088d51fcb49",
                        "createdAt": "27/04/2023 às 10:49:29"
                      },
                      {
                        "_id": "644a7db70ff85088d51fcb52",
                        "exercise": "Air Squat",
                        "sets": "3",
                        "reps": "8",
                        "type": "Legs",
                        "createdBy": "644a7d0d0ff85088d51fcb49",
                        "createdAt": "27/04/2023 às 10:50:47"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "There aren't exercises available. No user registered exercises.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Message"
                },
                "examples": {
                  "error": {
                    "value": {
                      "message": "No exercises avaible."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/exercises/exercise": {
      "post": {
        "description": "Route to create an enxercise.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exercise"
              },
              "examples": {
                "exercise": {
                  "value": {
                    "exercise": "Pull up",
                    "sets": "5",
                    "reps": "10",
                    "type": "pull"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request was successful. The exercise was created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "Exercise created."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error due to missing required fields provided by the client side.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "miss exercise field": {
                    "value": {
                      "message": "The following fields are required: exercise."
                    }
                  },
                  "miss sets field": {
                    "value": {
                      "message": "The following fields are required: sets."
                    }
                  },
                  "miss reps field": {
                    "value": {
                      "message": "The following fields are required: reps."
                    }
                  },
                  "miss type fields": {
                    "value": {
                      "message": "The following fields are required: type."
                    }
                  },
                  "miss all fields": {
                    "value": {
                      "message": "The following fields are required: exercise, sets, reps, type."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/exercises/exercise/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "schema": {
            "type": "string"
          },
          "required": true,
          "description": "The exercise ID."
        }
      ],
      "get": {
        "description": "Route to get an exercise by id.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                },
                "examples": {
                  "exercises": {
                    "value": {
                      "_id": "644aac5a0756b40273d69ac0",
                      "exercise": "Australian Pull ups",
                      "sets": "5",
                      "reps": "15",
                      "type": "pull",
                      "createdBy": "644a7d0d0ff85088d51fcb49",
                      "createdAt": "23/04/2023 às 10:23:29",
                      "inWorkouts": []
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400-ObjectIdCheck"
          },
          "404": {
            "$ref": "#/components/responses/404-getExerciseByIdAndCheck"
          }
        }
      },
      "put": {
        "description": "Route to update an exercise by id.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/exerciseUpdateExamples"
        },
        "responses": {
          "200": {
            "description": "Exercise updated with success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Message"
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "Exercise updated."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Id passed as parameter is invalid or the exercise wasn't updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Message"
                },
                "examples": {
                  "error - invalid id": {
                    "$ref": "#/components/examples/invalid-id"
                  },
                  "error - not updated": {
                    "$ref": "#/components/examples/not-updated"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401-ifParamIdIsUserIdCheck"
          },
          "404": {
            "$ref": "#/components/responses/404-getExerciseByIdAndCheck"
          }
        }
      },
      "delete": {
        "description": "Route to delete an exercise by id.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise deleted with success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "The exercise was deleted with success."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Id passed as parameter is invalid or the exercise wasn't deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "error - invalid id": {
                    "$ref": "#/components/examples/invalid-id"
                  },
                  "error - not deleted": {
                    "$ref": "#/components/examples/not-deleted"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401-ifParamIdIsUserIdCheck"
          },
          "404": {
            "$ref": "#/components/responses/404-getExerciseByIdAndCheck"
          }
        }
      },
      "post": {
        "description": "Route to copy an exercise. It can be the user's own or another user's.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Copied with success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "message": {
                    "value": {
                      "message": "Copied with success."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400-ObjectIdCheck"
          },
          "404": {
            "$ref": "#/components/responses/404-getExerciseByIdAndCheck"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Exercise": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "exercise": {
            "type": "string"
          },
          "sets": {
            "type": "string"
          },
          "reps": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "copiedExerciseId": {
            "type": "string"
          },
          "inWorkouts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Exercises": {
        "type": "array",
        "items": {
          "type": "object",
          "$ref": "#/components/schemas/Exercise"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "400-ObjectIdCheck": {
        "description": "Id passed as parameter is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/Message"
            },
            "examples": {
              "error": {
                "value": {
                  "message": "Invalid Id."
                }
              }
            }
          }
        }
      },
      "404-getExerciseByIdAndCheck": {
        "description": "Exercise not found in database.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            },
            "examples": {
              "error": {
                "value": {
                  "message": "Exercise not found."
                }
              }
            }
          }
        }
      },
      "401-ifParamIdIsUserIdCheck": {
        "description": "Id passed as parameter not linked to the user.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            },
            "examples": {
              "error": {
                "value": {
                  "message": "This id is not linked to this user."
                }
              }
            }
          }
        }
      },
...
    },
    "requestBodies": {
      "exerciseUpdateExamples": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Exercise"
            },
            "examples": {
              "update": {
                "value": {
                  "exercise": "Deadlift",
                  "sets": "8",
                  "reps": "10",
                  "type": "Fullbody"
                }
              },
              "exercise-update": {
                "value": {
                  "exercise": "Deadlift"
                }
              },
              "sets-update": {
                "value": {
                  "sets": "8"
                }
              },
              "reps-update": {
                "value": {
                  "reps": "10"
                }
              },
              "type-update": {
                "value": {
                  "type": "Fullbody"
                }
              }
            }
          }
        }
      },
...
    },
    "examples": {
      "invalid-id": {
        "value": {
          "message": "Invalid Id."
        }
      },
      "not-updated": {
        "value": {
          "message": "Wasn't updated."
        }
      },
      "not-deleted": {
        "value": {
          "message": "Wasn't deleted."
        }
      },
...
    }
  }
}